move _get_example_expedition back to utils
