+++ tests/cli/test_initialise.py
@@ -1,0 +1,14 @@
+from virtualship.models.expedition import Expedition
+from virtualship.utils import _get_example_expedition
+
+
+def test_get_example_expedition():
+    assert len(_get_example_expedition()) > 0
+
+
+def test_valid_example_expedition(tmp_path):
+    path = tmp_path / "test.yaml"
+    with open(path, "w") as file:
+        file.write(_get_example_expedition())
+
+    Expedition.from_yaml(path)
